Moved header into the headers directory.  #370
