+++ apps/build_disk_index.cpp
@@ -9,7 +9,7 @@
 #include "math_utils.h"
 #include "index.h"
 #include "partition.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 
 namespace po = boost::program_options;
