+++ apps/build_memory_index.cpp
@@ -7,7 +7,7 @@
 
 #include "index.h"
 #include "utils.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 
 #ifndef _WINDOWS
 #include <sys/mman.h>
