+++ apps/range_search_disk_index.cpp
@@ -15,7 +15,7 @@
 #include "pq_flash_index.h"
 #include "partition.h"
 #include "timer.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 
 #ifndef _WINDOWS
 #include <sys/mman.h>
