+++ apps/search_memory_index.cpp
@@ -20,7 +20,7 @@
 #include "index.h"
 #include "memory_mapper.h"
 #include "utils.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 #include "index_factory.h"
 
 namespace po = boost::program_options;
