+++ apps/test_insert_deletes_consolidate.cpp
@@ -11,7 +11,7 @@
 #include <future>
 
 #include "utils.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 #include "index_factory.h"
 
 #ifndef _WINDOWS
