+++ apps/test_streaming_scenario.cpp
@@ -13,7 +13,7 @@
 #include <index_factory.h>
 
 #include "utils.h"
-#include "../src/program_options_utils.hpp"
+#include "../include/program_options_utils.hpp"
 
 #ifndef _WINDOWS
 #include <sys/mman.h>
